backend/requirements.txt
Django==4.2.23
gunicorn==23.0.0
djangorestframework==3.16.0
django-filter==25.1
psycopg2-binary==2.9.10
dj-database-url==3.0.1
django-tinymce==5.0.0
django-cors-headers==4.7.0
django-imagekit==5.0.0
django-cleanup
drf-spectacular
django-unfold
# force rebuild
Pillow==11.3.0
pillow-heif
whitenoise==6.9.0
python-dotenv==1.1.1
django-colorfield==0.9.0
pytils
requests==2.32.3
celery
redis
